Move reporting additional dates transform to correct table
